backend/requirements.txt
Django==3.1
djangorestframework==3.12.2
django-environ==0.4.5
django-extensions==3.1.3
psycopg2-binary>=2.8
